Agent Implementation
Basic Agent Functionality
Player Corrections
Decision Tree Beginning - Needs Editing
Removal of Test Tree
New Bullet and Rocket Types

+++ AI-Project-Implementation/Assets/Scripts/Interfaces/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void TakeDamage(int argAmount);
+    public void TakeDamage(float argAmount);
 }
